+++ package.json
@@ -29,7 +29,7 @@
     "copy-placeholder-config": "cp ./src/config/config.placeholder.ts ./src/config/index.ts",
     "test-pptr-headless": "HEADLESS=true jest --config=jest-puppeteer.config.js",
     "test-pptr": "JEST_PPTR_RETRY_TIMES=0 HEADLESS=false jest --config=jest-puppeteer.config.js",
-    "run-playwright": "yarn playwright test --workers=10",
+    "run-playwright": "npm run playwright test --workers=10",
     "test": "jest",
     "preview": "vite preview"
   },
@@ -74,7 +74,7 @@
     "@babel/preset-env": "7.21.5",
     "@babel/preset-react": "7.18.6",
     "@babel/preset-typescript": "7.21.5",
-    "@playwright/test": "^1.51.2",
+    "@playwright/test": "1.51.2",
     "@puppeteer/browsers": "2.2.1",
     "@swc-node/jest": "1.6.6",
     "@swc/core": "1.3.74",
