+++ playwright-tests/createNft/actions.ts
@@ -1,11 +1,17 @@
+import { Page } from '@playwright/test';
 import { DataTestIdsEnum } from '../../src/localConstants/dataTestIds.enum';
 import { GlobalDataEnum } from '../utils/enums';
 
+interface NftCreation {
+  page: Page;
+  colectionName?: string;
+  nftName?: string;
+}
 export const createNft = async ({
   page,
   colectionName = 'col',
   nftName = 'TestNft'
-}) => {
+}: NftCreation) => {
   await page.waitForTimeout(5000);
   await page.locator('.css-19bb58m').click();
   await page.getByRole('option', { name: colectionName }).first().click();
