+++ playwright-tests/send/Send-Nft.spec.ts
@@ -2,8 +2,9 @@ import { test } from '@playwright/test';
 import { sendAction } from './actions.ts';
 import { sendNftData } from './data.ts';
 import { DataTestIdsEnum } from '../../src/localConstants/dataTestIds.enum.ts';
-import { login, handlePopup } from '../utils/actions.ts';
 import { LoginFilesEnum } from '../utils/enums.ts';
+import { handlePopup } from '../utils/handlePopup.ts';
+import { login } from '../utils/login.ts';
 
 test('Positive Send Nft', async ({ page }) => {
   await page.goto('/unlock');
