+++ playwright-tests/utils/confirmPem.ts
@@ -1,0 +1,16 @@
+import { Page } from '@playwright/test';
+import { DataTestIdsEnum } from '../../src/localConstants/dataTestIds.enum';
+import { LoginFilesEnum } from '../utils/enums.ts';
+
+interface ConfirmPem {
+  page: Page;
+  file?: string;
+}
+export const confirmPem = async ({
+  page,
+  file = LoginFilesEnum.pem
+}: ConfirmPem) => {
+  await page.getByText('Select a file').click();
+  await page.setInputFiles(DataTestIdsEnum.inputFile, file);
+  await page.getByTestId(DataTestIdsEnum.submitButton).click();
+};
